feat: upgrade to Angular 16, mark components as standalone
This commit upgrades the workspace to Angular 16, as it includes security fixes necessary
to ensure compatibility with content security policies. All components have been designated
as `standalone`, eliminating the need for them to be declared within the `HotToastModule`.
Consequently, `HotToastModule` is now deprecated, as it is unnecessary for standalone applications.